requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==19.3.0             # via pytest
click==7.1.2              # via -r requirements.in
importlib-metadata==1.6.1  # via pluggy, pytest
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest
pathtools==0.1.2          # via watchdog
pluggy==0.13.1            # via pytest
py==1.8.2                 # via pytest
pyparsing==2.4.7          # via packaging
pytest==5.4.3             # via -r requirements.in
pyyaml==5.3.1             # via -r requirements.in
six==1.15.0               # via packaging
watchdog==0.10.2          # via -r requirements.in
wcwidth==0.2.4            # via pytest
zipp==3.1.0               # via importlib-metadata
